refactor: 타입, 상숫값 파일로 분리
